requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
attrs==25.4.0
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
distro==1.9.0
Django==5.2.3
ezdxf==1.1.0
fonttools==4.58.5
fpdf==1.7.2
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
kiwisolver==1.4.8
lxml==6.0.0
matplotlib==3.10.3
multidict==6.7.1
mysqlclient==2.2.7
numpy==2.3.1
openai==2.21.0
packaging==25.0
pillow==11.2.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
razorpay==2.0.1
reportlab==4.4.2
requests==2.32.5
rsa==4.9.1
scipy==1.16.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
